Share macro code.
